tree-wide: refactor python scripts
