+++ .gitignore
@@ -1,1 +1,3 @@
 local.sh
+
+*.pyc
