+++ reporter.py
@@ -1,5 +1,3 @@
-#!/usr/local/bin/python
-
 import io
 import json
 import sys
@@ -7,8 +5,11 @@ import os
 import requests
 import time
 import signal
+import argparse
 
 from dwq import Disque, Job
+
+from common import parse_job
 
 
 sys.stdout = io.TextIOWrapper(sys.stdout.detach(), "utf-8", "replace")
@@ -23,28 +24,8 @@ signal.signal(signal.SIGTERM, signal_handler)
 signal.signal(signal.SIGINT, signal_handler)
 
 
-def has_passed(job):
-    if not job:
-        return False
-    return job["result"]["status"] in { 0, "0", "pass" }
-
-
-def job_name(job):
-    return os.path.join(*job["result"]["body"]["command"].split()[1:])
-
-
-def job_result_filename(job, cwd=None):
-    jobname = job_name(job) + ".txt"
-    filename = os.path.join("output", jobname)
-    cwd = cwd or os.getcwd()
-    if os.path.abspath(filename).startswith(cwd):
-        return filename
-    else:
-        return None
-
-
 def save_job_result(job):
-    filename = job_result_filename(job)
+    filename = os.path.join("output", f"{job['name']}.txt")
     if filename:
         os.makedirs(os.path.dirname(filename), exist_ok=True)
         with open(filename, "w") as f:
@@ -52,13 +33,7 @@ def save_job_result(job):
         return filename
 
 
-def dump_to_file(path, data):
-    with open(path + ".tmp", "w") as f:
-        f.write(data)
-    os.rename(path + ".tmp", path)
-
-
-def update_status(data, uid, failed_jobs, failed_builds, failed_tests, http_root):
+def update_status(data, uid, token, failed_jobs, failed_builds, failed_tests, http_root):
     http_root = os.path.join("/", http_root)
     status = {}
     # copy expected (but optional) fields that are in data
@@ -70,49 +45,35 @@ def update_status(data, uid, failed_jobs, failed_builds, failed_tests, http_root
     # "href" [optional]) field
     if failed_jobs:
         status["failed_jobs"] = []
-        for filename, jobname, _, _, _, _ in failed_jobs:
+        for jobname in failed_jobs:
             failed_job = {"name": jobname}
-            if filename:
-                failed_job["href"] = os.path.join(http_root, filename)
             status["failed_jobs"].append(failed_job)
 
     if failed_builds:
         status["failed_builds"] = []
-        for filename, jobname, board, toolchain, worker, runtime  in failed_builds:
+        for application, board, toolchain, worker, runtime  in failed_builds:
             failed_build = {
-                "name": jobname.replace("compile/", ""),
+                "application": application,
                 "board": board,
                 "toolchain": toolchain,
                 "worker": worker,
                 "runtime": runtime,
             }
-            if filename:
-                failed_build["href"] = os.path.join(http_root, filename)
             status["failed_builds"].append(failed_build)
 
     if failed_tests:
         status["failed_tests"] = []
-        for filename, jobname, board, toolchain, worker, runtime  in failed_tests:
+        for application, board, toolchain, worker, runtime  in failed_tests:
             failed_test = {
-                "name": jobname.replace("run_test/", ""),
+                "application": application,
                 "board": board,
                 "toolchain": toolchain,
                 "worker": worker,
                 "runtime": runtime,
             }
-            if filename:
-                failed_test["href"] = os.path.join(http_root, filename)
             status["failed_tests"].append(failed_test)
 
-    do_put_status(status, uid)
-
-
-def do_put_status(status, uid):
-    token = sys.argv[3]
     data = json.dumps({"uid" : uid, "status" : status})
-
-    dump_to_file("prstatus.json", json.dumps(status, indent=True))
-
     requests.put(
         f'http://localhost:8000/jobs/running/{uid}/status',
         headers={"Authorization": token},
@@ -121,13 +82,19 @@ def do_put_status(status, uid):
 
 
 def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("queue", type=str, help="Name of the queue to listen to")
+    parser.add_argument("job_uid", type=str, help="UID of the Murdock job")
+    parser.add_argument("job_token", type=str, help="Authentication token of the Murdock job")
+    args = parser.parse_args()
+    queue = args.queue
+    uid = args.job_uid
+    token = args.job_token
+
     disque_url = os.environ.get("DWQ_DISQUE_URL", "localhost:7711")
     Disque.connect([disque_url])
 
     http_root = os.environ.get("CI_BUILD_HTTP_ROOT", "/")
-
-    queue = sys.argv[1]
-    uid = sys.argv[2]
 
     last_update = 0
 
@@ -149,31 +116,29 @@ def main():
             job = _status.get('job')
 
             if job:
+                result = parse_job(job)
                 filename = save_job_result(job)
-                result = job["result"]
 
-                if filename and not has_passed(job):
-                    jobname = job_name(job)
-                    board = result["board"] if "board" in result else ""
-                    toolchain = result["toolchain"] if "toolchain" in result else ""
+                if filename and result["status"] is False:
+                    jobname = result["name"]
                     worker = result["worker"]
                     runtime = result["runtime"]
                     if jobname == "static_tests":
                         nfailed_jobs += 1
-                        failed_jobs = [(filename, jobname, "", "", worker, runtime)] + failed_jobs
+                        failed_jobs.append(jobname)
 
                     elif jobname.startswith("compile/"):
                         nfailed_builds += 1
                         if nfailed_builds <= maxfailed_builds:
                             failed_builds.append(
-                                (filename, job_name(job), board, toolchain, worker, runtime)
+                                (result["application"], result["board"], result["toolchain"], worker, runtime)
                             )
 
                     elif jobname.startswith("run_test/"):
                         nfailed_tests += 1
                         if nfailed_tests <= maxfailed_tests:
                             failed_tests.append(
-                                (filename, job_name(job), board, toolchain, worker, runtime)
+                                (result["application"], result["board"], result["toolchain"], worker, runtime)
                             )
 
                     failed_jobs = failed_jobs[:maxfailed_jobs]
@@ -195,13 +160,9 @@ def main():
 
             now = time.time()
             if now - last_update > 0.5:
-                update_status(_status, uid, failed_jobs, failed_builds, failed_tests, http_root)
+                update_status(_status, uid, token, failed_jobs, failed_builds, failed_tests, http_root)
                 last_update = now
 
 
 if __name__=="__main__":
-    argc = len(sys.argv)
-    if argc != 4:
-        print(f"error: {sys.argv[0]} <queue-name> <job uid> <job token>", file=sys.stderr)
-        sys.exit(1)
     main()
